删掉重复的测试，合并利用相同表的测试
