+++ start-db-core/src/test/scala/org/urbcomp/start/db/ddl/DatabaseExecutorTest.scala
@@ -9,26 +9,14 @@
  * General Public License for more details.
  */
 
-package org.urbcomp.start.db.executor
+package org.urbcomp.start.db.ddl
 
 import org.junit.Assert.{assertFalse, assertTrue}
 import org.urbcomp.start.db.AbstractCalciteFunctionTest
 
 class DatabaseExecutorTest extends AbstractCalciteFunctionTest {
 
-  test("test create then show databases") {
-    val stmt = connect.createStatement()
-    val databaseName = "test_%d".format(scala.util.Random.nextInt(100000))
-    stmt.executeUpdate("CREATE DATABASE %s".format(databaseName))
-    val rs = stmt.executeQuery("SHOW DATABASES")
-    var databases = List[String]()
-    while (rs.next()) {
-      databases = databases :+ rs.getString(1)
-    }
-    assertTrue(databases.contains(databaseName))
-  }
-
-  test("test create then drop database") {
+  test("test create, show and drop database") {
     val stmt = connect.createStatement()
     val databaseName = "test_%d".format(scala.util.Random.nextInt(100000))
 
@@ -53,7 +41,7 @@ class DatabaseExecutorTest extends AbstractCalciteFunctionTest {
     val stmt = connect.createStatement()
     val databaseName = "test_%d".format(scala.util.Random.nextInt(100000))
 
-    stmt.executeUpdate("CREATE DATABASE %s".format(databaseName))
+    stmt.executeUpdate("CREATE DATABASE IF NOT EXISTS %s".format(databaseName))
     val rs1 = stmt.executeQuery("SHOW DATABASES")
     var databasesBefore = List[String]()
     while (rs1.next()) {
