+++ start-db-core/src/test/scala/org/urbcomp/start/db/ddl/TableExecutorTest.scala
@@ -9,7 +9,7 @@
  * General Public License for more details.
  */
 
-package org.urbcomp.start.db.executor
+package org.urbcomp.start.db.ddl
 
 import org.junit.Assert.{assertEquals, assertNotNull}
 import org.urbcomp.start.db.AbstractCalciteFunctionTest
@@ -19,7 +19,7 @@ import org.urbcomp.start.db.model.trajectory.Trajectory
 import scala.collection.mutable.ArrayBuffer
 
 class TableExecutorTest extends AbstractCalciteFunctionTest {
-  test("test create table") {
+  test("test create, show and drop table") {
     val randomNum = scala.util.Random.nextInt(100000)
     val createTableSQL = s"""CREATE TABLE start_db_table_test_%d (
                                   |    idx Integer,
@@ -45,44 +45,17 @@ class TableExecutorTest extends AbstractCalciteFunctionTest {
                                   |);""".format(randomNum).stripMargin
     val stmt = connect.createStatement()
     stmt.executeUpdate(createTableSQL)
-  }
-
-  test("test show tables") {
-    val stmt = connect.createStatement()
+    // test show tables
     val rs = stmt.executeQuery("show tables")
     rs.next()
     assertNotNull(rs.getString(1))
+    // test drop table
+    val dropTableSQL = s"""DROP TABLE start_db_table_test_%d;""".format(randomNum).stripMargin
+    val cnt = stmt.executeUpdate(dropTableSQL)
+    assertEquals(1, cnt)
+
   }
 
-  test("test drop table") {
-    val randomNum = scala.util.Random.nextInt(100000)
-    val createTableSQL = s"""CREATE TABLE xxx_%d (
-                            |    idx Integer,
-                            |    ride_id String,
-                            |    x1 String,
-                            |    x2 String
-                            |);""".format(randomNum).stripMargin
-    val stmt = connect.createStatement()
-
-    val rs1 = stmt.executeQuery("show tables")
-    val tablesBefore = ArrayBuffer[String]()
-    while (rs1.next()) {
-      tablesBefore += rs1.getString(1)
-    }
-
-    stmt.executeUpdate(createTableSQL)
-
-    val dropTableSQL = s"""DROP TABLE xxx_%d;""".format(randomNum).stripMargin
-    stmt.executeUpdate(dropTableSQL)
-
-    val rs2 = stmt.executeQuery("show tables")
-    val tablesAfter = ArrayBuffer[String]()
-    while (rs2.next()) {
-      tablesAfter += rs2.getString(1)
-    }
-
-    assertEquals(tablesBefore.sorted, tablesAfter.sorted)
-  }
 
   test("test drop table if exists") {
     val randomNum = scala.util.Random.nextInt(100000)
@@ -114,25 +87,8 @@ class TableExecutorTest extends AbstractCalciteFunctionTest {
     assertEquals(tablesBefore.sorted, tablesAfter.sorted)
   }
 
-  test("test describe table") {
+  test("test describe table and show create table") {
     val tableName = "test_describe_table_" + scala.util.Random.nextInt(Integer.MAX_VALUE)
-    val createTableSQL = s"""CREATE TABLE $tableName (
-                            |    tr Trajectory,
-                            |    rs RoadSegment,
-                            |    gm Geometry
-                            |);""".stripMargin
-    val stmt = connect.createStatement()
-    stmt.executeUpdate(createTableSQL)
-    val rss = stmt.executeQuery(s"describe $tableName")
-    var fields = List[String]()
-    while (rss.next()) {
-      fields = fields :+ s"${rss.getString(1)}:${rss.getString(2)}:${rss.getString(3)}"
-    }
-    assertEquals(3, fields.length)
-  }
-
-  test("test show create table") {
-    val tableName = "test_show_create_table"
     val createTableSQL = s"""CREATE TABLE IF NOT EXISTS $tableName (
                             |    tr Trajectory,
                             |    rs RoadSegment,
@@ -140,16 +96,27 @@ class TableExecutorTest extends AbstractCalciteFunctionTest {
                             |);""".stripMargin
     val stmt = connect.createStatement()
     stmt.executeUpdate(createTableSQL)
-    val rss = stmt.executeQuery(s"show create table $tableName")
-    if (!rss.next()) {
+    // test for show create table
+    val rs = stmt.executeQuery(s"show create table $tableName")
+    if (!rs.next()) {
       throw new AssertionError("unexpected show create table no result");
     }
-    val sql = rss.getString(2);
+    val sql = rs.getString(2);
     assertEquals(
-      "CREATE TABLE test_show_create_table (tr Trajectory, rs RoadSegment, gm Geometry)",
+      s"CREATE TABLE $tableName (tr Trajectory, rs RoadSegment, gm Geometry)",
       sql
     )
+    // test for create table
+    val rss = stmt.executeQuery(s"describe $tableName")
+    var fields = List[String]()
+    while (rss.next()) {
+      fields = fields :+ s"${rss.getString(1)}:${rss.getString(2)}:${rss.getString(3)}"
+    }
+    assertEquals(3, fields.length)
+    // drop table
+    stmt.executeUpdate(s"drop table if exists $tableName")
   }
+
 
   test("test truncate table") {
     val trajectory: Trajectory = ModelGenerator.generateTrajectory()
@@ -185,5 +152,7 @@ class TableExecutorTest extends AbstractCalciteFunctionTest {
       resultSize2 = resultSize2 + 1
     }
     assertEquals(resultSize2, 1)
+    // drop table
+    stmt.executeUpdate(s"drop table if exists $tableName")
   }
 }
